% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_call.R
\name{format_call}
\alias{format_call}
\title{to fill in}
\usage{
format_call(call = NULL, verbose = FALSE, ...)
}
\arguments{
\item{call}{to fill in}

\item{verbose}{to fill in}
}
\description{
to fill in
}
\details{
to fill in
}
\examples{
format_call(call = "harvestR::get_table(table = 'time_entries', user = HARVEST_USER, key = HARVEST_KEY, query = list(from ='2020-07-01', to = '2020-07-15'))")

}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
