% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_call.R
\name{run_call}
\alias{run_call}
\title{Execute R call from string argument}
\usage{
run_call(call = NULL, env_vars = NULL, libs = NULL, verbose = TRUE, ...)
}
\arguments{
\item{call}{R call}

\item{libs}{R libs}

\item{verbose}{logical scalar}

\item{env_Vars}{creds and dynamic params}
}
\description{
to fill in
}
\details{
to fill in
}
\examples{
run_call(call = "harvestR::get_table(table = 'time_entries', user = HARVEST_USER, key = HARVEST_KEY, query = list(from ='2020-07-01', to = '2020-07-15'), strategy = \'multiprocess\')")

}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
